Ensure object existence check before deletion.
Previously, the `deleteObject` method was called without verifying the object's existence, potentially causing errors. Added a check using `objectExists` to confirm the object's presence before attempting deletion. This update enhances the robustness of the deletion process.

+++ src/persistence/ObjectsAutoSaver.cjs
@@ -316,7 +316,9 @@ function AutoSaverPersistence(storageStrategy, periodicInterval) {
         await this.deleteIndexedField(id, typeName);
         delete cache[id];
         delete modified[id];
-        await storageStrategy.deleteObject(id);
+        if(await storageStrategy.objectExists(id)){
+            await storageStrategy.deleteObject(id);
+        }
     }
     this.removeFromGrouping = async function (typeName, objId) {
         let obj = await loadWithCache(objId);
